Refactor supervisor to receive flags
Previously, we had separate configurations for the standard/debug
versions of each service.  Now, the $FLAGS environment variable is
sent to the services that require options instead.

Supervisor cannot handle non-existent environment variables, so
you *have* to specify FLAGS, even if it is empty.  This is addressed
in the Makefile, by modifying the supervisord script to be `FLAGS=$(FLAGS)
supervisord` instead, so that it reuses any existing FLAGS definition,
or defines it as empty otherwise.

This change will necessitate modifications to Cesium and others
relying on baselayer.