Correctly route OAuth2 requests to FakeOAuth2 server
Previously, we tried to verify OAuth2 credentials by posting to the
same URL as used by the browser + `/fakeoauth2/token`.  The token
request is made internally (from the web browser), resulting in
problems when hosting via Docker.

For example, the browser may point to `http://localhost:9000` because
Docker remaps 5000 -> 9000.  But from inside the Docker
container (where the server runs), the FakeOAuth URL remains
`http://localhost:5000/fakeoauth2` (not 9000).

The workaround here is to connect directly to the hosted port of the
service (default 63000).