Allow connection to databases with passwords.
Previously, we relied on `pg_hba.conf` to be configured so that no
password was required for db connection.  Technically, baselayer already
supported connections with passwords, but the script that tested the db
connection (`tools/db_init.py`) did not, causing early failure.

`psql` is still provided with the `--no-password` flag, indicating that
we never want the user to enter the password by hand.  But now, the
`PGPASSWORD` environment variable is set correctly if a password is
found in the configuration file.

Closes skyportal/skyportal#103