Fix supervisorctl monitoring of micro-services
Supervisor 4 introduces an error code on exited services
(https://github.com/Supervisor/supervisor/issues/1223).

The internal function that calls supervisorctl has been modified to
return the output as well as the error code. When the error code is 3,
we now proceed as though the operation was successful (which, typicall,
it was).