Ensure OAuth UID matches username (#142)
* Ensure OAuth UID matches username

Currently, if a server gets re-deployed with a new database, previously
logged in users are able to access the system.  Often, their UIDs now
point to a new user, so they "log in" as this new person.

To address this, we now, in addition to user_id, set a cookie for OAuth
UID (Google email). When logging in, we check this cookie to
ensure that the user doesn't get logged into a different username.

* Ignore cases where SocialAuth is not present on User

This can happen when users are generated by script, e.g.