Add migration manager
The migration manager checks to see whether migrations are present.
If so, it attempts to perform a migration if the DB is not up to date
already.  It then runs an HTTP server to report the current migration
status.

This status can be queried by the app factory to decide whether to
launch or not.